We have some real results. Many things matter. Right now we estimate bandwidth every 20 iterations, otherwise this add too much varaiance. M-estimators also make a fairly big difference. Better a inital values also matter a lot.
Right now the outstanding issue is how to pick c. Its a bit large often
currenlty. Small values with large ranges means that we push agaist the boundry a bit.

Thinking of maxamizing a \bar{x}/\sigma parameter. Or at least ensuring its no silly.
while keeping c as small as possible.

Next to do. Fix the code to not be a pile hack.